Initiated Receiver sketch
Someone says use the Arduino timer to convert data into PPM is not acurate, also the FC has to convert PPM back to data then PWM to motor, it might not work properly. For this sketch I use it mainly testing the TX signal and planned make it use for simulator.