Raised up RX battery minimum voltage and warning to protect my LiPo battery
Since I am using a LiPo battery package so it is not a good idea to discharge it until 3.2v and then give warning. 3.2v was for 18650 (it can be 3.0 - 3.2). The BMS module cut off at 2.9 according to the spec. I should've moved this to a global setting area for easy config for different types of battery.